+++ src/main/java/test/crudboard/controller/PostController.java
@@ -10,13 +10,14 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
-import test.crudboard.domain.aop.annotation.CheckResourceOwner;
+import test.crudboard.domain.aop.valid.CheckResourceOwner;
 import test.crudboard.domain.entity.post.Post;
 import test.crudboard.domain.entity.post.dto.CreatePostDto;
 import test.crudboard.domain.entity.post.dto.PostDetailDto;
-import test.crudboard.domain.entity.enumtype.ResourceType;
+import test.crudboard.domain.aop.valid.ResourceType;
 import test.crudboard.security.provider.JwtUserDetails;
 import test.crudboard.service.LikeService;
+import test.crudboard.service.PostHeaderService;
 import test.crudboard.service.PostService;
 import test.crudboard.service.RedisService;
 
@@ -31,12 +32,15 @@ import test.crudboard.service.RedisService;
 public class PostController {
     private final PostService postService;
     private final LikeService likeService;
-    private final RedisService redisService;
+    private final PostHeaderService postHeaderService;
 
     @PreAuthorize("isAuthenticated()")
     @GetMapping("/create")
     public String createPage(@AuthenticationPrincipal JwtUserDetails user, Model model){
+
+        //게시글 작성칸, 작성자를 표시하기 위해 현재 로그인한 사용자의 이름을 모델에 반환
         model.addAttribute("createPostDto",new CreatePostDto(user.getUsername()));
+
         return "post";
     }
 
@@ -45,10 +49,14 @@ public class PostController {
     public String createPost(@AuthenticationPrincipal JwtUserDetails userDetails,
                              @Valid @ModelAttribute CreatePostDto createPostDto,
                              BindingResult bindingResult){
+        //유효성 검사 실패 시 수정 페이지로 다시 이동
         if(bindingResult.hasErrors()){
             return "post";
         }
+
+        //문제없다면 저장
         postService.save(createPostDto, userDetails.getId());
+
         return "redirect:/";
     }
 
@@ -57,7 +65,9 @@ public class PostController {
                                 @RequestParam(value = "isrecommend", required = false, defaultValue = "false") boolean isRecommend,
                                 @AuthenticationPrincipal JwtUserDetails user, Model model){
 
-        PostDetailDto dto = postService.getPostDetailDtoById(postId, isRecommend);
+        //게시글, 댓글 부분 각각 불러옴
+        PostDetailDto dto = postHeaderService.getPostDetailDtoById(postId, isRecommend);
+
         model.addAttribute("header", dto.getHeader());
         model.addAttribute("footer", dto.getCommentList());
         model.addAttribute("currentUserNickname", user != null ? user.getUsername() : null);
@@ -92,6 +102,7 @@ public class PostController {
                            @Valid @ModelAttribute CreatePostDto postDto,
                            BindingResult bindingResult,
                            Model model) {
+
         // 유효성 검사 실패 시 수정 페이지로 다시 이동
         if (bindingResult.hasErrors()) {
             model.addAttribute("postId", postId);
@@ -105,6 +116,8 @@ public class PostController {
     @PreAuthorize("isAuthenticated()")
     @PostMapping("/recommend/{postId}")
     public String recommendPost(@PathVariable Long postId, @AuthenticationPrincipal JwtUserDetails user){
+
+        //추천, 이미 추천했다면 적용되지 않음
         likeService.recommendPost(postId,user.getId());
 
         return "redirect:/post/" + postId;
@@ -115,6 +128,7 @@ public class PostController {
     @CheckResourceOwner(type = ResourceType.POST)
     @ResponseBody
     public void deletePost(@PathVariable Long id, @AuthenticationPrincipal JwtUserDetails user){
+        //게시글 삭제
         postService.deletePost(id);
     }
 }
