+++ src/main/java/test/crudboard/domain/aop/valid/ResourceType.java
@@ -1,0 +1,7 @@
+package test.crudboard.domain.aop.valid;
+
+
+//사용자 검증시 사용
+public enum ResourceType {
+    POST, COMMENT
+}
