+++ src/main/java/test/crudboard/domain/entity/comment/Comment.java
@@ -59,19 +59,24 @@ public class Comment {
         return this.isParent;
     }
 
+    //양방향 연관관계 설정
     public void setPost(Post post) {
         this.post = post;
         post.getCommentList().add(this);
     }
 
+    //양방향 연관관계 설정
     public void setUser(User user) {
         this.user = user;
         user.getCommentList().add(this);
     }
 
+    //이름이 setparent와 중복되어 에러가 발생했기에, 따로 수정
     public void setIsParent(boolean isParent){
         this.isParent = isParent;
     }
+
+    //양방향 연관관계 설정
 
     public void addChild(Comment comment) {
         if (!comment.isParent) {
@@ -80,9 +85,14 @@ public class Comment {
         }else System.out.println("!comment.isParent error");
     }
 
+
+    /**
+     *  jpa를 통한 엔티티를 저장하기 위해, 연관관계 설정을 위해 필요한 id 필드만 저장해 새 객체 생성
+     */
     private Comment(Long id){
         this.id = id;
     }
+
 
     public static Comment Quick(Long id){
         return new Comment(id);
