+++ src/main/java/test/crudboard/domain/entity/comment/dto/PostFooterDto.java
@@ -9,6 +9,9 @@ import test.crudboard.domain.entity.comment.Comment;
 import java.util.List;
 
 
+/**
+ *  댓글을 저장하기 위한 dto
+ */
 @Getter
 @Setter
 @AllArgsConstructor
