--- src/main/java/test/crudboard/domain/entity/enumtype/AuthProvider.java
@@ -1,5 +1,0 @@
-package test.crudboard.domain.entity.enumtype;
-
-public enum AuthProvider {
-    LOCAL,GITHUB
-}
