--- src/main/java/test/crudboard/domain/entity/enumtype/ResourceType.java
@@ -1,5 +1,0 @@
-package test.crudboard.domain.entity.enumtype;
-
-public enum ResourceType {
-    POST, COMMENT
-}
