--- src/main/java/test/crudboard/domain/entity/post/dto/MainTitleDto.java
@@ -1,17 +1,0 @@
-package test.crudboard.domain.entity.post.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import java.time.LocalDateTime;
-
-@Data
-@AllArgsConstructor
-public class MainTitleDto {
-    private Long id;
-    private String head;
-    private String name;
-    private Long view;
-    private LocalDateTime created;
-    private Long likeCount;
-    private Long commentCount;
-}
