+++ src/main/java/test/crudboard/domain/entity/post/dto/PostHeader.java
@@ -11,6 +11,9 @@ import org.springframework.data.redis.core.TimeToLive;
 import java.time.LocalDateTime;
 
 
+/**
+ * redis 에 게시글 정보를 저장하기 위한 객체
+ */
 @Getter
 @Setter
 @NoArgsConstructor
@@ -27,7 +30,8 @@ public class PostHeader {
     private Long comment_count;
 
     private String nickname;
-    @Setter
+
+    //ttl(캐시 생존 시간)설정 위한 변수
     @TimeToLive
     private Long ttl;
