+++ src/main/java/test/crudboard/domain/entity/post/dto/SearchRequestDto.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 import test.crudboard.domain.entity.enumtype.SearchType;
 
+
+//검색 데이터를 받는 dto
 @Data
 public class SearchRequestDto {
     private SearchType type;
