+++ src/main/java/test/crudboard/domain/entity/user/User.java
@@ -6,8 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import test.crudboard.domain.entity.comment.Comment;
-import test.crudboard.domain.entity.enumtype.AuthProvider;
-import test.crudboard.domain.entity.enumtype.Roles;
+import test.crudboard.security.type.AuthProvider;
+import test.crudboard.security.type.Roles;
 import test.crudboard.domain.entity.like.Like;
 import test.crudboard.domain.entity.post.Post;
