+++ src/main/java/test/crudboard/repository/JpaCommentRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import test.crudboard.domain.entity.comment.Comment;
-import test.crudboard.domain.entity.comment.dto.CommentPageDto;
+
 
 import java.util.List;
 
@@ -24,20 +24,6 @@ public interface JpaCommentRepository extends JpaRepository<Comment, Long> {
     @Query("DELETE FROM Comment c WHERE c.id = :commentId")
     void deleteCommentById(@Param("commentId") Long commentId);
 
-    @Query("select new test.crudboard.domain.entity.comment.dto.CommentPageDto" +
-            "(c.id, " +         //comment
-            "c.content, " +
-            "c.parent.id, " +
-            "c.isParent, " +
-            "c.depth, " +
-            "c.created, "+
-            "c.post.id, " +     //post
-            "c.user.id, " +     //user
-            "c.user.nickname)" +
-            " from Comment c " +
-            "where c.post.id = :postId " +
-            "order by coalesce(c.parent.id, c.id) , c.isParent desc , c.id asc")
-    Page<CommentPageDto> findCommentPageByPostId(@Param("postId") Long postId, Pageable pageRequest);
 
     long countByPostId(Long postId);
