+++ src/main/java/test/crudboard/security/provider/JwtUserDetails.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
-import test.crudboard.domain.entity.enumtype.Roles;
+import test.crudboard.security.type.Roles;
 
 import java.util.Collection;
 import java.util.Collections;
