+++ src/main/java/test/crudboard/security/type/AuthProvider.java
@@ -1,0 +1,7 @@
+package test.crudboard.security.type;
+
+
+// 로그인한 사용자 플랫폼 구분
+public enum AuthProvider {
+    LOCAL,GITHUB
+}
