+++ src/main/java/test/crudboard/security/type/Roles.java
@@ -1,4 +1,4 @@
-package test.crudboard.domain.entity.enumtype;
+package test.crudboard.security.type;
 
 public enum Roles {
     ROLE_USER,
