+++ src/main/java/test/crudboard/service/LogoutService.java
@@ -6,6 +6,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
@@ -22,10 +23,15 @@ public class LogoutService {
      * @param request
      */
     public void invalidateToken(HttpServletRequest request){
+        //엑세스 토큰 추출
         String token = jwtService.extractToken(request);
         if(token == null){
-            return;
+            throw new BadCredentialsException("잘못된 접근");
         }
+
+        /**
+         * 해당 토큰의 유효기간을 확인해서, 아직 유효기간이 남아있다면, 그 시간만큼 캐시에서 블랙리스트로 관리
+         */
         Claims claims = jwtService.parseClaims(token);
         Date expiration = claims.getExpiration();
         long ttl = (expiration.getTime() - System.currentTimeMillis()) / 1000;
